pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "trafilatura>=2.0.0",
]
